LayoutTests/accessibility/aria-label-expected.txt

This tests that the aria-label attribute works on both an input and an anchor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS platformValueForW3CName(axInput) is "aria label"
PASS platformValueForW3CName(axLink) is "aria link"
PASS successfullyParsed is true

TEST COMPLETE
